refactor(visualizer): useTheme hook
