+++ packages/visualizer/src/index.tsx
@@ -23,6 +23,7 @@ export type { NavActionsProps } from './component/nav-actions';
 
 export { Logo } from './component/logo';
 export { iconForStatus, timeCostStrElement } from './component/misc';
+export { useTheme } from './hooks/useTheme';
 export { useServerValid } from './hooks/useServerValid';
 export {
   useSafeOverrideAIConfig,
